+++ src/handlers/engines/readability.ts
@@ -1,7 +1,7 @@
 import { Readability } from '@mozilla/readability';
-import { HandlerInput } from './handler-input';
-import { IHandlerOutput } from './handler.interface';
-import { EngineParseError } from '../errors/main';
+import { HandlerInput } from '../handler-input';
+import { IHandlerOutput } from '../handler.interface';
+import { EngineParseError } from '../../errors/main';
 
 export default async function readability(
   input: HandlerInput
